requirements.txt
gcloud==0.18.3
GDAL==2.2.3
googleapis-common-protos==1.5.2
httplib2==0.10.3
jws==0.1.3
numpy==1.16.2
oauth2client==4.1.2
protobuf==3.3.0
psycopg2==2.7.7
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodome==3.6.6
PyMySQL==0.7.11
Pyrebase4==4.3.0
python-jwt==2.0.1
requests==2.20.0
requests-toolbelt==0.7.1
rocketchat-API==0.6.5
rsa==3.4.2
six==1.10.0
